Using "%*c" instead of "%*s" for padding saves a byte
The saving is in the second argument, which can be 0 instead of ""

We do have some one-letter variables that could also be used as
arguments, but they would result in various unwanted output (in the
last line, since the end is not overwritten for non-multiples of 16).

The character will be left-padded to the end of the line buffer, and
having a null byte at the end of lines is what we want anyway.